Refactored shuttle-specific code into ShuttleMapFragment from MitMapFragment
