+++ app/src/main/java/edu/mit/mitmobile2/MitMapFragment.java
@@ -1,13 +1,9 @@
 package edu.mit.mitmobile2;
 
 import android.app.Fragment;
-import android.app.LoaderManager;
-import android.content.Loader;
 import android.content.res.Configuration;
-import android.database.Cursor;
 import android.location.Location;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
@@ -20,7 +16,6 @@ import android.widget.ArrayAdapter;
 import android.widget.FrameLayout;
 import android.widget.ListView;
 import android.widget.RelativeLayout;
-import android.widget.TextView;
 
 import com.getbase.floatingactionbutton.FloatingActionButton;
 import com.google.android.gms.maps.CameraUpdate;
@@ -29,60 +24,41 @@ import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.MapView;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
-import com.google.android.gms.maps.model.PolylineOptions;
 
 import java.util.ArrayList;
-import java.util.List;
-import java.util.Timer;
-import java.util.TimerTask;
 
 import edu.mit.mitmobile2.maps.MITMapView;
 import edu.mit.mitmobile2.maps.MapItem;
-import edu.mit.mitmobile2.shuttles.model.MITShuttleRoute;
 
-public abstract class MitMapFragment extends Fragment implements Animation.AnimationListener, LoaderManager.LoaderCallbacks<Cursor>, GoogleMap.OnMapLoadedCallback {
-    private static final int PREDICTIONS_PERIOD = 15000;
-    private static final int PREDICTIONS_TIMER_OFFSET = 1000;
-
+public abstract class MitMapFragment extends Fragment implements Animation.AnimationListener, GoogleMap.OnMapLoadedCallback, GoogleMap.InfoWindowAdapter {
     public static final int NO_TRANSLATION = 0;
     public static final int ANIMATION_LENGTH = 500;
 
-    public static final int STOP_ZOOM = 17;
-
     private static final String MAP_EXPANDED_KEY = "mapExpanded";
 
-    private ListView mapItemsListView;
-    private ArrayList mapItems;
+    protected ListView mapItemsListView;
     protected MITMapView mitMapView;
 
-    private View transparentView;
-    private View transparentLandscapeView;
+    protected View transparentView;
+    protected View transparentLandscapeView;
 
-    private boolean hasHeader = false;
-    private RelativeLayout routeInfoSegment;
+    protected RelativeLayout headerInfoSegment;
     protected SwipeRefreshLayout swipeRefreshLayout;
+    private boolean swipeRefreshEnabled = false;
 
     protected boolean mapViewExpanded = false;
-    private boolean animating = false;
+    protected boolean animating = false;
 
-    private boolean stopMode = false;
-    private FrameLayout shuttleStopContent;
-    //Stores the offset done by showing the stop in the header (or left side for landscape)
-    protected double latOffset;
-    protected double lonOffset;
+    protected boolean stopMode = false;
+    protected FrameLayout shuttleStopContent;
 
     private int originalPosition = -1;
-
-    private static Timer timer;
 
     private FloatingActionButton listButton;
     private FloatingActionButton myLocationButton;
 
-    protected boolean isRoutePredictable = false;
-    protected boolean isRouteScheduled = false;
+    protected int headerLayout = R.layout.default_transparent_header;
 
-    @Nullable
-    @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.activity_solo_map, null);
 
@@ -118,10 +94,12 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
             myLocationButton.setVisibility(View.VISIBLE);
         }
 
+        addHeaderView(View.inflate(getActivity(), headerLayout, null));
+
         return view;
     }
 
-    private void setupFabButtons() {
+    protected void setupFabButtons() {
         listButton.setSize(FloatingActionButton.SIZE_NORMAL);
         listButton.setColorNormalResId(R.color.mit_red);
         listButton.setColorPressedResId(R.color.mit_red_dark);
@@ -151,45 +129,19 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
         });
     }
 
-    private void setupMapView() {
+    protected void setupMapView() {
         mitMapView.setMapViewExpanded(mapViewExpanded);
         if (!mapViewExpanded) {
             getMapView().getUiSettings().setAllGesturesEnabled(false);
         }
 
-        mitMapView.getMap().setInfoWindowAdapter(new GoogleMap.InfoWindowAdapter() {
-            @Override
-            public View getInfoWindow(Marker marker) {
-                return null;
-            }
-
-            @Override
-            public View getInfoContents(Marker marker) {
-                if (marker.getTitle() != null) {
-                    View view = View.inflate(getActivity(), R.layout.mit_map_info_window, null);
-                    TextView stopNameTextView = (TextView) view.findViewById(R.id.top_textview);
-                    TextView stopPredictionView = (TextView) view.findViewById(R.id.bottom_textview);
-                    stopNameTextView.setText(marker.getTitle());
-                    if (isRoutePredictable) {
-                        stopPredictionView.setText(marker.getSnippet());
-                    } else if (isRouteScheduled) {
-                        stopPredictionView.setText(getString(R.string.route_unknown));
-                    } else {
-                        stopPredictionView.setText(getString(R.string.route_not_in_service));
-                    }
-                    return view;
-                } else {
-                    return null;
-                }
-            }
-        });
+        mitMapView.getMap().setInfoWindowAdapter(this);
     }
 
     //Setup for shuttle route
     protected void addHeaderView(View headerView) {
-        hasHeader = true;
         mapItemsListView.addHeaderView(headerView);
-        routeInfoSegment = (RelativeLayout) headerView.findViewById(R.id.route_info_segment);
+        headerInfoSegment = (RelativeLayout) headerView.findViewById(R.id.route_info_segment);
         transparentView = headerView.findViewById(R.id.transparent_map_overlay);
 
         if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
@@ -204,7 +156,7 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
                 }
             });
         }
-        swipeRefreshLayout.setEnabled(true);
+        swipeRefreshLayout.setEnabled(swipeRefreshEnabled);
         swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
@@ -224,7 +176,7 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
                 public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
                     int topRowPosition = (mapItemsListView == null || mapItemsListView.getChildCount() == 0) ?
                             0 : mapItemsListView.getChildAt(0).getTop();
-                    swipeRefreshLayout.setEnabled((topRowPosition >= 0));
+                    swipeRefreshLayout.setEnabled((topRowPosition >= 0) && swipeRefreshEnabled);
 
                     if (!mapViewExpanded && !animating) {
                         int newPosition = calculateScrollOffset();
@@ -258,7 +210,7 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
                 public void onScroll(AbsListView view, int firstVisibleItem, int visibleItemCount, int totalItemCount) {
                     int topRowPosition = (mapItemsListView == null || mapItemsListView.getChildCount() == 0) ?
                             0 : mapItemsListView.getChildAt(0).getTop();
-                    swipeRefreshLayout.setEnabled((topRowPosition >= 0));
+                    swipeRefreshLayout.setEnabled((topRowPosition >= 0) && swipeRefreshEnabled);
                 }
             });
         }
@@ -269,32 +221,6 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
                 listItemClicked(position);
             }
         });
-    }
-
-    //Setup for shuttle stop
-    protected void addShuttleStopContent(View content) {
-        stopMode = true;
-        shuttleStopContent.addView(content);
-        swipeRefreshLayout.setVisibility(View.GONE);
-        shuttleStopContent.setVisibility(View.VISIBLE);
-
-        if (transparentView != null) {
-            transparentView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    if (!animating) {
-                        toggleMap();
-                    }
-                }
-            });
-        } else {
-            transparentLandscapeView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    toggleMapHorizontal();
-                }
-            });
-        }
     }
 
     protected void addTransparentView(View view) {
@@ -313,24 +239,6 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
         mapItemsListView.setAdapter(arrayAdapter);
     }
 
-    protected void drawRoutePath(MITShuttleRoute route) {
-        for (List<List<Double>> outerList : route.getPath().getSegments()) {
-            List<LatLng> points = new ArrayList<>();
-            PolylineOptions options = new PolylineOptions();
-            for (List<Double> innerList : outerList) {
-                LatLng point = new LatLng(innerList.get(1), innerList.get(0));
-                points.add(point);
-            }
-            options.addAll(points);
-            options.color(getResources().getColor(R.color.map_path_color));
-            options.visible(true);
-            options.width(12f);
-            options.zIndex(100);
-
-            getMapView().addPolyline(options);
-        }
-    }
-
     public void showListView() {
         if (!animating) {
             listButton.setVisibility(View.INVISIBLE);
@@ -343,7 +251,7 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
         }
     }
 
-    private void toggleMap() {
+    protected void toggleMap() {
         DisplayMetrics displayMetrics = getResources().getDisplayMetrics();
 
         TranslateAnimation translateAnimation;
@@ -393,7 +301,7 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
         }
     }
 
-    private void toggleMapHorizontal() {
+    protected void toggleMapHorizontal() {
         DisplayMetrics displayMetrics = getResources().getDisplayMetrics();
 
         TranslateAnimation translateAnimation;
@@ -442,8 +350,8 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
 
     private int calculateScrollOffset() {
         int[] location = new int[2];
-        if (hasHeader) {
-            routeInfoSegment.getLocationOnScreen(location);
+        if (headerInfoSegment != null) {
+            headerInfoSegment.getLocationOnScreen(location);
         }
         return location[1];
     }
@@ -469,21 +377,16 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
         return mapViewExpanded;
     }
 
-    private void startTimerTask() {
-        timer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                updateData();
-            }
-        }, PREDICTIONS_TIMER_OFFSET, PREDICTIONS_PERIOD);
-    }
-
     protected void updateData() {
 
     }
 
     protected void queryDatabase() {
 
+    }
+
+    public void setSwipeRefreshEnabled(boolean swipeRefreshEnabled) {
+        this.swipeRefreshEnabled = swipeRefreshEnabled;
     }
 
     @Override
@@ -513,25 +416,7 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
     }
 
     @Override
-    public Loader<Cursor> onCreateLoader(int id, Bundle args) {
-        return null;
-    }
-
-    @Override
-    public void onLoadFinished(Loader<Cursor> loader, Cursor data) {
-
-    }
-
-    @Override
-    public void onLoaderReset(Loader<Cursor> loader) {
-
-    }
-
-    @Override
     public void onPause() {
-        timer.cancel();
-        timer.purge();
-        timer = null;
         mitMapView.getGoogleMapView().onPause();
         super.onPause();
     }
@@ -540,9 +425,6 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
     public void onResume() {
         super.onResume();
         mitMapView.getGoogleMapView().onResume();
-        timer = new Timer();
-        queryDatabase();
-        startTimerTask();
     }
 
     @Override
@@ -573,5 +455,15 @@ public abstract class MitMapFragment extends Fragment implements Animation.Anima
                 mitMapView.adjustCameraToShowInHeader(false, 0, getActivity().getResources().getConfiguration().orientation);
             }
         }
+    }
+
+    @Override
+    public View getInfoWindow(Marker marker) {
+        return null;
+    }
+
+    @Override
+    public View getInfoContents(Marker marker) {
+        return null;
     }
 }
