+++ app/src/main/res/layout/default_transparent_header.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <View
+        android:id="@+id/transparent_map_overlay"
+        android:layout_width="fill_parent"
+        android:layout_height="@dimen/shuttle_routes_map_header_height"
+        android:background="@color/transparent" />
+
+    <RelativeLayout
+        android:id="@+id/route_info_segment"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"/>
+</LinearLayout>
