+++ app/src/main/res/menu/mitmodule.xml
@@ -27,11 +27,4 @@ xmlns:mitmobile="http://schemas.android.com/apk/res-auto" >
     mitmobile:actionViewClass="android.support.v7.widget.SearchView"
     android:title="Search"/>
 
-<item
-    android:id="@+id/action_settings"
-    android:orderInCategory="100"
-    android:showAsAction="never"
-    mitmobile:showAsAction="never"
-    android:title="@string/action_settings"/>
-
 </menu>
